requirements.txt
appdirs==1.4.4
autopep8==1.5.7
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
chromedriver-binary==91.0.4472.19.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
distlib==0.3.1
filelock==3.0.12
flake8==3.9.2
future==0.18.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
httplib2==0.19.1
idna==2.10
importlib-metadata==2.1.1
importlib-resources==5.1.2
lxml==4.6.3
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.1.0
pipenv==2020.11.15
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.17.1
python-twitter==3.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
tweepy==3.10.0
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.4.6
virtualenv-clone==0.5.4
webdriver-manager==3.4.1
zipp==3.4.1
mysqlclient==*
